rename initialData to sectionData for clarity
